order_model_modified_order_fetch

+++ lib/pages/admin/orders_admin/widgets/single_order.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 class SingleOrder extends StatelessWidget {
   final String orderName;
-  final int orderAmount;
+  final double orderAmount;
   final String orderDate;
   const SingleOrder({super.key,required this.orderName, required this.orderAmount, required this.orderDate});
